move const data to ground event
